+++ lexio/lib/components/ContentDisplay/ContentDisplay.tsx
@@ -2,7 +2,7 @@ import React, { useContext } from 'react';
 import { PdfViewer } from "../Viewers/PdfViewer";
 import { HtmlViewer } from "../Viewers/HtmlViewer";
 import { MarkdownViewer } from "../Viewers/MarkdownViewer";
-import { useRAGSources } from "../RAGProvider/hooks";
+import { useRAGSources } from "../../hooks/hooks";
 import { ThemeContext, removeUndefined } from "../../theme/ThemeContext";
 
 export interface ContentDisplayStyles extends React.CSSProperties {
