+++ lexio/lib/components/RAGProvider/index.tsx
@@ -1,5 +1,3 @@
 import { RAGProvider } from './provider';
-import { useRAGStatus } from './hooks';
-import { useLexio, useRAGMessages, useRAGSources } from './hooks';
 
-export { RAGProvider, useRAGStatus, useLexio, useRAGMessages, useRAGSources };
+export { RAGProvider };
